pyproject.toml
[project]
name = "magic-shop"
version = "0.1.0"
description = "A whimsical fictional magic item store with AI-generated products using Google Gemini"
requires-python = ">=3.11"
authors = [
    { name = "Magic Shop Team" }
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "jinja2>=3.1.2",
    "python-multipart>=0.0.6",
    "google-genai>=0.2.0",
    "pillow>=10.1.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests that require real API credentials (deselect with '-m \"not integration\"')"
]
